+++ src/KingmakerGunslinger/Blueprints/GunslingerClassBlueprints.cs
@@ -31,7 +31,8 @@ namespace KingmakerGunslinger.Blueprints
             ExpertLoadingBlueprintSet expertLoading,
             LightningReloadBlueprintSet lightningReload,
             EvasiveBlueprintSet evasive, MenacingShotBlueprintSet menacingShot,
-            SlingersLuckBlueprintSet slingersLuck, BlueprintFeature cheatDeath)
+            SlingersLuckBlueprintSet slingersLuck, BlueprintFeature cheatDeath,
+            StunningShotBlueprintSet stunningShot)
         {
             CharacterClass = characterClass ?? throw new ArgumentNullException("characterClass");
             Progression = progression ?? throw new ArgumentNullException("progression");
@@ -66,6 +67,8 @@ namespace KingmakerGunslinger.Blueprints
             SlingersLuck = slingersLuck ??
                 throw new ArgumentNullException("slingersLuck");
             CheatDeath = cheatDeath ?? throw new ArgumentNullException("cheatDeath");
+            StunningShot = stunningShot ??
+                throw new ArgumentNullException("stunningShot");
         }
         internal BlueprintCharacterClass CharacterClass { get; private set; }
         internal BlueprintProgression Progression { get; private set; }
@@ -91,7 +94,8 @@ namespace KingmakerGunslinger.Blueprints
         internal MenacingShotBlueprintSet MenacingShot { get; private set; }
         internal SlingersLuckBlueprintSet SlingersLuck { get; private set; }
         internal BlueprintFeature CheatDeath { get; private set; }
-        internal int Count { get { return 5 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count + ExpertLoading.Count + LightningReload.Count + Evasive.Count + MenacingShot.Count + SlingersLuck.Count; } }
+        internal StunningShotBlueprintSet StunningShot { get; private set; }
+        internal int Count { get { return 5 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count + ExpertLoading.Count + LightningReload.Count + Evasive.Count + MenacingShot.Count + SlingersLuck.Count + StunningShot.Count; } }
     }
 
     internal sealed class GunslingerClassCatalogPublication
@@ -206,6 +210,8 @@ namespace KingmakerGunslinger.Blueprints
                     characterClass);
             BlueprintFeature cheatDeath = CheatDeathBlueprints.Register(registry,
                 grit.Resource, characterClass);
+            StunningShotBlueprintSet stunningShot = StunningShotBlueprints.Register(
+                library, registry, grit.Resource, characterClass);
             BlueprintProgression progression = registry.Register<BlueprintProgression>(
                 ProgressionSymbol, () => CreateProgression());
 
@@ -232,6 +238,7 @@ namespace KingmakerGunslinger.Blueprints
             progression.LevelEntries[14].Features.Add(menacingShot.Feature);
             progression.LevelEntries[14].Features.Add(slingersLuck.Feature);
             progression.LevelEntries[18].Features.Add(cheatDeath);
+            progression.LevelEntries[18].Features.Add(stunningShot.Feature);
             Validate(characterClass, progression, proficiencies, fullBab, goodSave,
                 poorSave, startingPistol, blackPowder, leadBall,
                 simple, martial, lightArmor, firearmProficiency);
@@ -240,7 +247,7 @@ namespace KingmakerGunslinger.Blueprints
                 pistolWhip, utilityShot, gunTraining, deadShot, startlingShot,
                 targetingHead, targetingTorso, targetingLegs, bleedingWound,
                 expertLoading, lightningReload, evasive, menacingShot,
-                slingersLuck, cheatDeath);
+                slingersLuck, cheatDeath, stunningShot);
         }
 
         internal static GunslingerClassCatalogPublication Publish(